Agents/env_generator/generated/calendar/calendar_api/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.30.0
SQLAlchemy~=2.0.0
pydantic[email]~=2.9.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.0
psycopg2-binary~=2.9.0